TitleScreen/MainMenu, Options Menu, Pause Functional
Title or main screen menu whatever is functional, hitting escape in game will show the pause menu AND pause the game but NOT the pause menu. The buttons on these menu's are functional. The OPTIONS menu from the main screen is navigatable but doesn't have anything in it yet, thinking about adding an easter egg in it to where it'll disable having to be on the floor for jumping (flight lmao)
Added sounds to a good amount of stuff.